Updater and Conditions
This update allows row updates with conditions to be expressed in Java
code.  Use cases include the following.

Atomic updates.  A row can be optimistically locked via one of its
columns.  An update which contains the current column state as a
condition will fail if the row has been modified in the meantime.

Furthermore, mass updates to tables can be expressed as we allow column
updates, basic column math, and a vast set of conditions.  This feature
set however will require some further testing.